+++ Application/ApiContracts/Output/Requests/UpdateOutputInfoRequest.cs
@@ -6,6 +6,6 @@ namespace Application.ApiContracts.Output.Requests
 
         public int? ProductId { get; set; }
 
-        public short? Count { get; set; }
+        public int? Count { get; set; }
     }
 }
