+++ Application/ApiContracts/Product/Common/VariantRow.cs
@@ -8,7 +8,7 @@
 
         public string? UrlSlug { get; set; }
 
-        public long? Price { get; set; }
+        public decimal? Price { get; set; }
 
         public string? CoverImageUrl { get; set; }
 
@@ -16,8 +16,8 @@
 
         public List<OptionPair> OptionPairs { get; set; } = [];
 
-        public long Stock { get; set; }
+        public int Stock { get; set; }
 
-        public long HasBeenBooked { get; set; }
+        public int HasBeenBooked { get; set; }
     }
 }
