+++ Application/ApiContracts/Product/Requests/UpdateProductPriceRequest.cs
@@ -2,5 +2,5 @@ namespace Application.ApiContracts.Product.Requests;
 
 public class UpdateProductPriceRequest
 {
-    public long? Price { get; set; }
+    public decimal? Price { get; set; }
 }
