+++ Application/ApiContracts/Product/Requests/UpdateVariantPriceRequest.cs
@@ -2,5 +2,5 @@ namespace Application.ApiContracts.Product.Requests;
 
 public class UpdateVariantPriceRequest
 {
-    public long? Price { get; set; }
+    public decimal? Price { get; set; }
 }
