+++ Application/ApiContracts/Statistical/Responses/DashboardStatsResponse.cs
@@ -2,12 +2,12 @@ namespace Application.ApiContracts.Statistical.Responses
 {
     public class DashboardStatsResponse
     {
-        public long LastMonthRevenue { get; set; }
+        public decimal LastMonthRevenue { get; set; }
 
-        public long LastMonthProfit { get; set; }
+        public decimal LastMonthProfit { get; set; }
 
-        public long PendingOrdersCount { get; set; }
+        public int PendingOrdersCount { get; set; }
 
-        public long NewCustomersCount { get; set; }
+        public int NewCustomersCount { get; set; }
     }
 }
