+++ Application/Features/Products/Commands/UpdateManyProductPrices/UpdateManyProductPricesCommand.cs
@@ -2,4 +2,4 @@ using MediatR;
 
 namespace Application.Features.Products.Commands.UpdateManyProductPrices;
 
-public sealed record UpdateManyProductPricesCommand(List<int> Ids, long Price) : IRequest<(List<int>? Data, Common.Models.ErrorResponse? Error)>;
+public sealed record UpdateManyProductPricesCommand(List<int> Ids, decimal Price) : IRequest<(List<int>? Data, Common.Models.ErrorResponse? Error)>;
