+++ Application/Features/Products/Commands/UpdateManyVariantPrices/UpdateManyVariantPricesCommand.cs
@@ -2,4 +2,4 @@ using MediatR;
 
 namespace Application.Features.Products.Commands.UpdateManyVariantPrices;
 
-public sealed record UpdateManyVariantPricesCommand(List<int> Ids, long Price) : IRequest<(List<int>? Data, Common.Models.ErrorResponse? Error)>;
+public sealed record UpdateManyVariantPricesCommand(List<int> Ids, decimal Price) : IRequest<(List<int>? Data, Common.Models.ErrorResponse? Error)>;
