+++ Application/Features/Products/Commands/UpdateVariantPrice/UpdateVariantPriceCommand.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace Application.Features.Products.Commands.UpdateVariantPrice;
 
-public sealed record UpdateVariantPriceCommand(int VariantId, long Price) : IRequest<(ProductVariantLiteResponse? Data, Common.Models.ErrorResponse? Error)>;
+public sealed record UpdateVariantPriceCommand(int VariantId, decimal Price) : IRequest<(ProductVariantLiteResponse? Data, Common.Models.ErrorResponse? Error)>;
