+++ Application/Features/Settings/Queries/GetAllSettings/GetAllSettingsQuery.cs
@@ -2,4 +2,4 @@ using MediatR;
 
 namespace Application.Features.Settings.Queries.GetAllSettings;
 
-public sealed record GetAllSettingsQuery : IRequest<Dictionary<string, long?>>;
+public sealed record GetAllSettingsQuery : IRequest<Dictionary<string, decimal?>>;
